+++ examples/tsoulos_nlode3.cpp
@@ -1,5 +1,5 @@
-#include <iostream>
 #include <audi/audi.hpp>
+#include <iostream>
 
 #include "../include/expression.hpp"
 #include "../include/kernel_set.hpp"
@@ -7,21 +7,22 @@
 // Here we solve the differential equation d^2y dy  = - 4 / x^3 (NLODE3) from Tsoulos paper
 // Tsoulos and Lagaris: "Solving Differential equations with genetic programming"
 
-double fitness(const dcgp::expression<gdual_d>& ex, const std::vector<std::vector<gdual_d> >& in)
+double fitness(const dcgp::expression<gdual_d> &ex, const std::vector<std::vector<gdual_d>> &in)
 {
     double retval = 0;
     for (auto i = 0u; i < in.size(); ++i) {
-        auto T = ex(in[i]);                   // We compute the expression and thus the derivatives
+        auto T = ex(in[i]); // We compute the expression and thus the derivatives
         double dy = T[0].get_derivative({1});
         double ddy = T[0].get_derivative({2});
         double x = in[i][0].constant_cf();
-        double ode1 = - 4 / x / x / x;
-        retval += (ode1 - ddy*dy) * (ode1 - ddy*dy);    // We compute the quadratic error
+        double ode1 = -4 / x / x / x;
+        retval += (ode1 - ddy * dy) * (ode1 - ddy * dy); // We compute the quadratic error
     }
     return retval;
 }
 
-int main () {
+int main()
+{
     // Random seed
     std::random_device rd;
 
@@ -35,34 +36,35 @@ int main () {
     std::vector<std::string> in_sym({"x"});
 
     // We create the grid over x
-    std::vector<std::vector<gdual_d> > in(10u);
+    std::vector<std::vector<gdual_d>> in(10u);
     for (auto i = 0u; i < in.size(); ++i) {
-        in[i].push_back(gdual_d(1. + 1. / (in.size() - 1u) * i, "x", 2)); // 1, .., 2
+        in[i].push_back(gdual_d(1. + 1. / static_cast<double>((in.size() - 1)) * i, "x", 2)); // 1, .., 2
     }
 
     // We run the (1-4)-ES
     double best_fit = 1e32;
     std::vector<double> newfits(4, 0.);
-    std::vector<std::vector<unsigned int> > newchromosomes(4);
+    std::vector<std::vector<unsigned int>> newchromosomes(4);
     std::vector<unsigned int> best_chromosome(ex.get());
     unsigned int gen = 0;
 
-    do
-    {
+    do {
         gen++;
         for (auto i = 0u; i < newfits.size(); ++i) {
             ex.set(best_chromosome);
             ex.mutate_active(2);
-            auto fitness_ic = ex(std::vector<gdual_d>{gdual_d(1.)})[0];        // Penalty term to enforce the initial conditions
-            newfits[i] = fitness(ex, in) + fitness_ic.constant_cf() * fitness_ic.constant_cf();     // Total fitness
+            auto fitness_ic
+                = ex(std::vector<gdual_d>{gdual_d(1.)})[0]; // Penalty term to enforce the initial conditions
+            newfits[i] = fitness(ex, in) + fitness_ic.constant_cf() * fitness_ic.constant_cf(); // Total fitness
             newchromosomes[i] = ex.get();
         }
 
         for (auto i = 0u; i < newfits.size(); ++i) {
             if (newfits[i] <= best_fit) {
                 if (newfits[i] != best_fit) {
-                    std::cout << "New best found: gen: " << std::setw(7) << gen << "\t value: " << newfits[i] << std::endl;
-                    //std::cout << "Expression: " << ex(in_sym) << std::endl;
+                    std::cout << "New best found: gen: " << std::setw(7) << gen << "\t value: " << newfits[i]
+                              << std::endl;
+                    // std::cout << "Expression: " << ex(in_sym) << std::endl;
                 }
                 best_fit = newfits[i];
                 best_chromosome = newchromosomes[i];
