+++ examples/tsoulos_ode1.cpp
@@ -6,21 +6,22 @@
 // Here we solve the differential equation dy = (2x - y) / x from Tsoulos paper
 // Tsoulos and Lagaris: "Solving Differential equations with genetic programming"
 
-double fitness(const dcgp::expression<gdual_d>& ex, const std::vector<std::vector<gdual_d> >& in)
+double fitness(const dcgp::expression<gdual_d> &ex, const std::vector<std::vector<gdual_d>> &in)
 {
     double retval = 0;
     for (auto i = 0u; i < in.size(); ++i) {
-        auto T = ex(in[i]);                   // We compute the expression and thus the derivatives
+        auto T = ex(in[i]); // We compute the expression and thus the derivatives
         double y = T[0].get_derivative({0});
         double dy = T[0].get_derivative({1});
         double x = in[i][0].constant_cf();
         double ode1 = (2. * x - y) / x;
-        retval += (ode1 - dy) * (ode1 - dy);            // We compute the quadratic error
+        retval += (ode1 - dy) * (ode1 - dy); // We compute the quadratic error
     }
     return retval;
 }
 
-int main () {
+int main()
+{
     // Random seed
     std::random_device rd;
 
@@ -34,10 +35,10 @@ int main () {
     std::vector<std::string> in_sym({"x"});
 
     // We create the grid over x
-    std::vector<std::vector<gdual_d> > in(10u);
+    std::vector<std::vector<gdual_d>> in(10u);
     for (auto i = 0u; i < in.size(); ++i) {
-        gdual_d point(0.1 + 0.9 / (in.size() - 1) * i, "x", 1);
-        in[i].push_back(point);   // 1, .., 2
+        gdual_d point(0.1 + 0.9 / static_cast<double>((in.size() - 1)) * i, "x", 1);
+        in[i].push_back(point); // 1, .., 2
     }
 
     // We run the (1-4)-ES
@@ -47,22 +48,22 @@ int main () {
     auto best_chromosome(ex.get());
     auto gen = 0u;
 
-    do
-    {
+    do {
         gen++;
         for (auto i = 0u; i < newfits.size(); ++i) {
             ex.set(best_chromosome);
             ex.mutate_active(2);
-            auto fitness_ic = ex({gdual_d(1.)})[0] - 3.;  // Penalty term to enforce the initial conditions
-            newfits[i] = fitness(ex, in) + fitness_ic.constant_cf() * fitness_ic.constant_cf();     // Total fitness
+            auto fitness_ic = ex({gdual_d(1.)})[0] - 3.; // Penalty term to enforce the initial conditions
+            newfits[i] = fitness(ex, in) + fitness_ic.constant_cf() * fitness_ic.constant_cf(); // Total fitness
             newchromosomes[i] = ex.get();
         }
 
         for (auto i = 0u; i < newfits.size(); ++i) {
             if (newfits[i] <= best_fit) {
                 if (newfits[i] != best_fit) {
-                    std::cout << "New best found: gen: " << std::setw(7) << gen << "\t value: " << newfits[i] << std::endl;
-                    //std::cout << "Expression: " << ex(in_sym) << std::endl;
+                    std::cout << "New best found: gen: " << std::setw(7) << gen << "\t value: " << newfits[i]
+                              << std::endl;
+                    // std::cout << "Expression: " << ex(in_sym) << std::endl;
                 }
                 best_fit = newfits[i];
                 best_chromosome = newchromosomes[i];
